Rename transcriber.py to aaf_writer.py
