+++ opentimelineio_contrib/adapters/aaf_adapter/aaf_writer.py
